Progress toward full rewrite
