+++ sec_edgar_downloader/_constants.py
@@ -1,0 +1,11 @@
+SEC_EDGAR_BASE_URL = "https://www.sec.gov/cgi-bin/browse-edgar?"
+SUPPORTED_FILINGS = {
+    "8-K",
+    "10-K",
+    "10-KSB",
+    "10-Q",
+    "13F-NT",
+    "13F-HR",
+    "SC 13G",
+    "SD",
+}
