+++ sec_edgar_downloader/_version.py
@@ -1,1 +1,1 @@
-__version__ = "2.2.1"
+__version__ = "3.0.0"
