+++ sec_edgar_downloader_old/__init__.py
@@ -1,0 +1,2 @@
+from ._version import __version__
+from .Downloader import Downloader
